Move logic to tools.py
